# Optimize WorkflowExecutor.propagateAllBlockedToReady