--- a/digdag-core/src/main/java/io/digdag/core/session/SessionStoreManager.java
+++ b/digdag-core/src/main/java/io/digdag/core/session/SessionStoreManager.java
@@ -47,6 +47,8 @@ public interface SessionStoreManager
     // for WorkflowExecutorManager.propagateSessionArchive
     List<TaskAttemptSummary> findRootTasksByStates(TaskStateCode[] states, long lastId);
 
+    List<Long> findDirectParentsOfBlockedTasks(long lastId);
+
     boolean requestCancelAttempt(long attemptId);
 
     int trySetRetryWaitingToReady();