--- a/digdag-core/src/main/java/io/digdag/core/session/SessionStoreManager.java
+++ b/digdag-core/src/main/java/io/digdag/core/session/SessionStoreManager.java
@@ -41,12 +41,13 @@ public interface SessionStoreManager
     // for WorkflowExecutorManager.IncrementalStatusPropagator.propagateStatus
     List<TaskStateSummary> findRecentlyChangedTasks(Instant updatedSince, long lastId);
 
-    // for WorkflowExecutorManager.propagateAllBlockedToReady
+    // for WorkflowExecutorManager.propagateAllPlannedToDone
     List<TaskStateSummary> findTasksByState(TaskStateCode state, long lastId);
 
     // for WorkflowExecutorManager.propagateSessionArchive
     List<TaskAttemptSummary> findRootTasksByStates(TaskStateCode[] states, long lastId);
 
+    // for WorkflowExecutorManager.propagateBlockedChildrenToReady
     List<Long> findDirectParentsOfBlockedTasks(long lastId);
 
     boolean requestCancelAttempt(long attemptId);