This method takes very long time if number of BLOCKED tasks
increased because it has a loop in it. Because the loop blocks
shutdown process, server returns 503 Service Unavailable to
clients for long time.

This method was fetching all BLOCKED tasks and get their parent_id.
If this method uses SQL to get the parent_id list directly, number of
the loop and data transfer from database to the server become several
times smaller.

```
=> select count(*) from (select id, parent_id, state from tasks where state = 0) orig;
  count
---------
 5278779
(1 row)

Time: 2877.191 ms
```

```
=> select count(*) from (select distinct parent_id from tasks where state = 0) optimized;
 count
--------
 135461
(1 row)

Time: 3972.704 ms
```

Query plan:

```
=> explain analyze select distinct parent_id from tasks where parent_id > 12345 and state = 0 order by parent_id limit 100;
                                                                            QUERY PLAN
-------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=0.43..356.94 rows=100 width=8) (actual time=0.039..3.068 rows=100 loops=1)
   ->  Unique  (cost=0.43..240934.82 rows=67581 width=8) (actual time=0.037..2.779 rows=100 loops=1)
         ->  Index Only Scan using tasks_on_parent_id_and_state on tasks  (cost=0.43..227809.86 rows=5249982 width=8) (actual time=0.033..1.553 rows=1088 loops=1)
               Index Cond: ((parent_id > 12345) AND (state = 0))
               Heap Fetches: 1023
 Planning time: 0.102 ms
 Execution time: 3.184 ms
(7 rows)
```

This query is scalable against number of tasks because query plan doesn't include SORT node.