@frsyuki 2016-09-19
I tried to combine the query with `trySetChildrenBlockedToReadyOrShortCircuitPlannedOrCanceled` but generated query becomes slow. It's better to avoid long-running single UPDATE query to not lock rows for long time.

```
explain
update tasks
set updated_at = now(), state = case
when task_type = 1 then 5
when (state_flags & 1) != 0 then 9
else 1
end
where parent_id = any(array(
  with recursive parents (id) as (
    (
      select parent_id as id
      from tasks
      where parent_id is not null and state = 0
      order by parent_id
      limit 1
    )
    union all
    select (
      select parent_id as id
      from tasks
      where parent_id > parents.id and state = 0
      order by parent_id
      limit 1
    )
    from parents
    where parents.id is not null
  )
  select parents.id
  from parents
  join tasks on parents.id = tasks.id
  where tasks.state in (5, 7)
))
and not exists (
  select * from tasks up
  join task_dependencies dep on up.id = dep.upstream_id
  where dep.downstream_id = tasks.id
  and up.parent_id = tasks.parent_id
  and up.state not in (7)
);
```